idk man

+++ blatt07/src/ringbuffer/main.cpp
@@ -1,5 +1,6 @@
 #include <iostream>
 #include <cstdlib>
+#include "ringbuffer.h"
 #include "CustomAllocator.h"
 
 using namespace std;
@@ -7,6 +8,8 @@ using namespace std;
 
 int main() {
     CustomAllocator<int> intAllocator;
-    intAllocator.allocate(5);
+
+    intAllocator.allocate((size_t)5);
+
     return EXIT_SUCCESS;
 }
